requirements.txt
pyserial>=3.5,<4
scipy
librosa
sounddevice
edge_impulse_linux
scikit-learn
edge_impulse_linux
numpy>=1.19,<3
PyAudio>=0.2.11,<0.3
six>=1.16.0,<2
opencv-python>=4.5.1.48,<5
matplotlib
python-dotenv>=1.0.0,<2
streamlit>=1.35,<2
fastapi>=0.110,<1
uvicorn[standard]>=0.20,<1
requests>=2.31,<3
